+++ Sources/SpotifyWebAPI/API/SpotifyAPI+Albums.swift
@@ -182,5 +182,4 @@ public extension SpotifyAPI {
         
     }
     
-    
 }