+++ Sources/SpotifyWebAPI/API/SpotifyAPI+Browse.swift
@@ -394,4 +394,3 @@ public extension SpotifyAPI {
     }
     
 }
-