+++ Sources/SpotifyWebAPI/Errors/SpotifyLocalError.swift
@@ -182,5 +182,3 @@ extension SpotifyLocalError: CustomStringConvertible {
     }
   
 }
-
-