+++ Sources/SpotifyWebAPI/Object Model/Audio Analysis Objects/AudioAnalysis.swift
@@ -127,5 +127,4 @@ extension AudioAnalysis: ApproximatelyEquatable {
 
     }
 
-
 }