+++ Sources/SpotifyWebAPI/Object Model/Other Spotify Objects/SpotifyCursor.swift
@@ -57,5 +57,3 @@ public struct SpotifyCursor: Codable, Hashable {
     }
 
 }
-
-