+++ Sources/SpotifyWebAPI/Object Model/Player Objects/CurrentlyPlayingContext.swift
@@ -349,5 +349,3 @@ extension CurrentlyPlayingContext: ApproximatelyEquatable {
         
     }
 }
-
-