+++ Sources/SpotifyWebAPI/Object Model/Post Request Objects/Playlist Objects/URIsDictWithInsertionIndex.swift
@@ -81,7 +81,6 @@ extension URIsDictWithInsertionIndex: Codable {
 
 }
 
-
 extension URIsDictWithInsertionIndex: Hashable {
     
     /// :nodoc: