+++ Sources/SpotifyWebAPI/Other/CodingExtensions.swift
@@ -212,7 +212,6 @@ public extension KeyedDecodingContainer {
     
 }
 
-
 public extension KeyedEncodingContainer {
     
     // MARK: - Spotify Dates -