+++ Sources/SpotifyWebAPI/Publisher Extensions/DelayedConditionalRetryPublisher.swift
@@ -210,5 +210,3 @@ extension Publisher {
     }
     
 }
-
-