+++ Tests/SpotifyWebAPIMainTests/Authorization Tests/SpotifyAPIAuthorizationCodeFlowPKCEAuthorizationTests.swift
@@ -43,7 +43,7 @@ final class SpotifyAPIAuthorizationCodeFlowPKCEAuthorizationTests:
             .sink(receiveValue: {
                 didDeauthorizeCount += 1
             })
-            .store(in: &Self.cancellables)
+            .store(in: &cancellables)
     
         let currentScopes = Self.spotify.authorizationManager.scopes ?? []
     
@@ -112,7 +112,7 @@ final class SpotifyAPIAuthorizationCodeFlowPKCEAuthorizationTests:
             .sink(receiveValue: {
                 didDeauthorizeCount += 1
             })
-            .store(in: &Self.cancellables)
+            .store(in: &cancellables)
         
         let currentAuthManager = Self.spotify.authorizationManager