+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.1] - 2021-4-15
+
+### Added
+
+* Added `SpotifyAPI.currentUserFollowedArtists(after:limit:)`.
+* Added documentation for each authorization scope that specifies which endpoints the scope is required for.
+* Added a ` makeCopy` method to `ClientCredentialsFlowManager`, `AuthorizationCodeFlowManager`, and `AuthorizationCodeFlowPKCEManager`.
+
 ## [1.5.0] - 2021-3-30
 
 ### Added